added utils.js to split code
